+++ js/util.js
@@ -1,0 +1,38 @@
+const getRandomPositiveInt = (a, b) => {
+  // реализуем поддержку передачи минимального и максимального значения в любом порядке,
+  // а какое из них большее и меньшее вычислим с помощью Math.min и Math.max.
+  // нижнюю границу диапазона мы округляем к ближайшему большему целому с помощью Math.ceil,
+  // а верхнюю границу - к ближайшему меньшему целому с помощью Math.floor
+  const lower = Math.ceil(Math.min(Math.abs(a), Math.abs(b)));
+  const upper = Math.floor(Math.max(Math.abs(a), Math.abs(b)));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+
+// Функция, возвращающая случайное число с плавающей точкой из переданного диапазона включительно.
+
+const getRandomPositiveFloat = (a, b, digits) => {
+  const lower = Math.min(Math.abs(a), Math.abs(b));
+  const upper = Math.max(Math.abs(a), Math.abs(b));
+  const result = Math.random() * (upper - lower) + lower;
+  return +result.toFixed(digits);
+};
+
+//Функция, которая перемешивает элементы массива
+
+const shuffleArray = (array) => {
+  for (let i = array.length - 1; i > 0; i--) {
+    const j = Math.floor(Math.random() * (i + 1));
+    [array[i], array[j]] = [array[j], array[i]];
+  }
+  return array;
+};
+
+const getImgNumber = (imgNumber) => {
+  if (imgNumber < 10) {
+    return `0${imgNumber}`;
+  }
+  return imgNumber;
+};
+
+export {getRandomPositiveInt, getRandomPositiveFloat, getImgNumber, shuffleArray};
